Clase 13/IdeaManager/requirements.txt
gradio==4.12.0
python-dotenv==1.0.0
openai==1.55.0
setuptools==67.6.1
